# app/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/empleado/": {
      "post": {
        "summary": "Create Empleado",
        "operationId": "create_empleado_empleado__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/empleado/{empleado_id}": {
      "get": {
        "summary": "Get Empleado",
        "operationId": "get_empleado_empleado__empleado_id__get",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update Empleado",
        "operationId": "update_empleado_empleado__empleado_id__post",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Empleado",
        "operationId": "delete_empleado_empleado__empleado_id__delete",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/empleados/": {
      "get": {
        "summary": "Get Empleados",
        "operationId": "get_empleados_empleados__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/perfil/{empleado_id}": {
      "post": {
        "summary": "Update Perfil",
        "operationId": "update_perfil_perfil__empleado_id__post",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Perfil",
        "operationId": "get_perfil_perfil__empleado_id__get",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Perfil",
        "operationId": "delete_perfil_perfil__empleado_id__delete",
        "parameters": [
          {
            "name": "empleado_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Empleado Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/perfiles/": {
      "get": {
        "summary": "Get Perfiles",
        "operationId": "get_perfiles_perfiles__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmpleadoBase": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "edad": {
            "type": "integer",
            "title": "Edad"
          },
          "correo": {
            "type": "string",
            "title": "Correo"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "responsabilidad": {
            "type": "string",
            "title": "Responsabilidad"
          }
        },
        "type": "object",
        "required": ["nombre", "edad", "correo", "rol", "responsabilidad"],
        "title": "EmpleadoBase"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PerfilBase": {
        "properties": {
          "empleado_id": {
            "type": "integer",
            "title": "Empleado Id"
          },
          "habilidad": {
            "type": "string",
            "title": "Habilidad"
          },
          "años_experiencia": {
            "type": "integer",
            "title": "Años Experiencia"
          },
          "certificacion": {
            "type": "string",
            "title": "Certificacion"
          },
          "tiempo_en_empresa": {
            "type": "integer",
            "title": "Tiempo En Empresa"
          },
          "salario": {
            "type": "integer",
            "title": "Salario"
          }
        },
        "type": "object",
        "required": [
          "empleado_id",
          "habilidad",
          "años_experiencia",
          "certificacion",
          "tiempo_en_empresa",
          "salario"
        ],
        "title": "PerfilBase"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}